JDP201101-23_1 rzeczywista implementacja CartController, problem z createOrder

+++ src/test/java/com/kodilla/ecommercee/OrderEntityTestSuite.java
@@ -154,9 +154,10 @@ public class OrderEntityTestSuite {
         int sizeListOfProductsFromOrder2 = order2.getProducts().size();
 
         long order3Id = order3.getOrderId();
-        orderRepository.deleteById((long) order3Id);
+        orderRepository.deleteById(order3Id);
         long order2Id = order2.getOrderId();
-        orderRepository.deleteById((long) order2Id);
+        orderRepository.deleteById(order2Id);
+
         long sizeOfOrderEntity = orderRepository.count();
         long sizeOfProductEntity = productRepository.count();
         long sizeOfUserEntity = userRepository.count();
